fix opacity

+++ src/prefs.h
@@ -41,7 +41,7 @@ struct _PREF
 	int winPosY;
 	int transparent;
 	int stealth;
-	int opacity;
+	float opacity;
 	GdkColor colors[16];
 	GdkColor fore[1];
 	GdkColor back[1];
